// src/app/models/item.model.ts
export class ItemModel {
  UniqueName: string;
  LocalizationNameVariable: string;
  LocalizationDescriptionVariable: string;
  LocalizedNames: {
    EN_US: string;
    DE_DE: string;
    FR_FR: string;
    RU_RU: string;
    PL_PL: string;
    ES_ES: string;
    PT_BR: string;
    ZH_CN: string;
    KO_KR: string;
  };
  LocalizedDescriptions: {
    EN_US: string;
    DE_DE: string;
    FR_FR: string;
    RU_RU: string;
    PL_PL: string;
    ES_ES: string;
    PT_BR: string;
    ZH_CN: string;
    KO_KR: string;
  };
  Index: string;
}